requirements.txt
annotated-types==0.5.0
apex==0.1
boto3==1.33.13
botocore==1.33.13
Bottleneck @ file:///opt/conda/conda-bld/bottleneck_1657175564434/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1725278078093/work/certifi
charset-normalizer==3.4.5
cycler==0.11.0
deepspeed==0.7.7
einops==0.6.1
filelock==3.12.2
fonttools==4.38.0
fsspec==2023.1.0
googledrivedownloader==0.4
hjson==3.1.0
huggingface-hub==0.16.4
idna==3.10
importlib-metadata==4.13.0
isodate==0.6.1
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.3.2
kiwisolver==1.4.5
Levenshtein==0.23.0
MarkupSafe==2.1.5
matplotlib==3.5.3
megatron-lm @ file:///workspace/MolBART/megatron_molbart/Megatron-LM-v1.1.5-3D_parallelism
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
msgpack==1.0.5
networkx==2.6.3
ninja==1.11.1.4
numexpr @ file:///croot/numexpr_1668713893690/work
numpy @ file:///opt/conda/conda-bld/numpy_and_numpy_base_1653915516269/work
nvidia-ml-py==13.590.48
ogb==1.2.0
packaging @ file:///croot/packaging_1671697413597/work
pandas==1.3.5
Pillow==9.4.0
psutil==7.2.2
py-cpuinfo==9.0.0
pybind11==2.13.6
pydantic==1.10.13
pydantic_core==2.14.6
pyparsing==3.1.4
PySMILESutils @ git+https://github.com/MolecularAI/pysmilesutils.git@b5b3b7d9755365d5a65de36693ef170c87db4e96
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz @ file:///croot/pytz_1671697431263/work
PyYAML==6.0.1
rapidfuzz==3.4.0
rdflib==6.3.2
regex==2024.4.16
requests==2.31.0
s3transfer==0.8.2
safetensors==0.3.0
scikit-learn==1.0.2
scipy==1.7.3
six @ file:///tmp/build/80754af9/six_1644875935023/work
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==1.9.1+cu111
torch-cluster==1.5.9
torch-geometric==2.0.3
torch-scatter==2.0.9
torch-sparse==0.6.12
torch-spline-conv==1.2.1
tqdm==4.67.3
transformers==4.24.0
typing_extensions==4.7.1
urllib3==1.26.20
yacs==0.1.8
zipp==3.15.0
